HAPP-1332: Create FetchImmunizationsUseCase
